Use fixed test domain in realmdomains test
Random domain name may bring undererministic behavior. It also breaks
the test on some systems as string.lowercase is locale dependent and
can return non-ASCII letters and thus later break the unicode encoding
and raise UnicodeDecodeError.

Use a fixed domain in "test" TLD instead. This domain is guaranteed to
be not existent.